refactor: rename to .ts & refactor indents in appLoader file
